requirements.txt
gdown==5.2.0
gpytorch==1.14
h5py==3.14.0
hydra-core==1.3.2
matplotlib==3.10.3
neuralop>=0.6.0
numpy==2.3.1
omegaconf==2.3.0
PyYAML==6.0.2
scikit_learn==1.7.0
scipy==1.15.3
torch==2.7.1
torchdiffeq==0.2.5
wandb>=0.15.0